+++ tests/common/mod.rs
@@ -1,4 +1,5 @@
 // Third-party
+use nalgebra::{Isometry3, Translation3, UnitQuaternion};
 use rand::RngExt;
 use rand_chacha::ChaCha8Rng;
 
@@ -19,6 +20,31 @@ pub fn assert_close(a: f64, b: f64) {
         (a - b).abs() < TEST_TOLERANCE,
         "expected {b}, got {a} OR not within {TEST_TOLERANCE}"
     );
+}
+
+/// Need to do this test because quaternions double-cover rotations (q=-q are same rotation)
+fn assert_orientation_close(a: &UnitQuaternion<f64>, b: &UnitQuaternion<f64>) {
+    let dot_prod = a.i * b.i + a.j * b.j + a.k * b.k + a.w * b.w;
+    assert_close(dot_prod.abs(), 1.0);
+}
+
+fn assert_position3d_close(a: &Translation3<f64>, b: &Translation3<f64>) {
+    assert_close(a.x, b.x);
+    assert_close(a.y, b.y);
+    assert_close(a.z, b.z);
+}
+
+pub fn assert_galaw_k_transform_close(galaw_transform: &Isometry3<f64>, k_iso: &k::nalgebra::Isometry3<f64>) {
+    assert_position3d_close(&galaw_transform.translation, &k_iso.translation);
+    assert_orientation_close(&galaw_transform.rotation, &k_iso.rotation);
+}
+
+/// Both sides are plain `nalgebra::Isometry3<f64>` here (dynamic vs codegen'd
+/// FK), unlike `assert_galaw_k_transform_close` which bridges to `k`'s own re-exported
+/// nalgebra type.
+pub fn assert_galaw_transform_close(a: &Isometry3<f64>, b: &Isometry3<f64>) {
+    assert_position3d_close(&a.translation, &b.translation);
+    assert_orientation_close(&a.rotation, &b.rotation);
 }
 
 /// Sets up the different kinematics model for testing.
